requirements.txt
pandas==1.4.4
numpy==1.21.5
requests==2.28.1
bs4==0.0.0
google-cloud-storage==2.7.0